refactor: simplify Expandable component structure by extracting subcomponents
